refactor(TableBoar): to import types from separate file
